Fixup tidy issues / broken test
Removes pybind11 headers from clang-tidy. This is not done automatically
because the cmake configuration for pybind11 does not tag the include
paths as system includes.

Further the change on the Simulator dummy is undone (change getValue()
to be a static function) because this will break the example system
tests. The corresponding tidy issues is locally silenced (This is just a
dummy after all)