Add clang-tidy to cmake
This commit brings a new configuration option for cmake WITH_TIDY.

If WITH_TIDY is set clang tidy will be enabeled to run alongside the
build (This is independent of the actual compiler used)

clang-tidy-10 is required exactly so that there is no discrepency
between the CI check and the check on your local machine.

WITH_TIDY is enabled by default.

A note about the find procedure for tidy:
This is untested on Windows and MacOS.
`clang-tidy` or `clang-tidy-10` binaries are accpected as long as the
`--version` output confirms it is verion 10.x