LengthUnit implementation (#38)
LengthUnit provides a strong type for all length unit quantities. It stores the actual unit using `double` precision floating points in the defined `RESOLUTION` (currently meter). It automatically scales the unit from the input unit to the internal storage and to the desired unit. 
We store the quantity in a single resolution unit to avoid a lot of scaling for arithmetic operations.