Python bindings for LengthUnit ctor and value access
Bindings for the constructor and the read only properties.
The constructor requires two arguments. The quantity and the unit of this quantity. Unfortunately we need a big switch statement which is matching the unit with a literal of the same unit to select the correct template function. Can be called in python like:

lu = LengthUnit(5.0, Units.mm)

The read only property call the getter of the length unit to retrieve the quantity in the desired unit. Can be called like:

print(lu.um)

Unfortunately a lot of copy and paste since templated functions can not be set by dynmic variables.