Bugfix for overwriting other events
When generating pedestrians the overwrite flag was checked twice. If
overwrite was set, the events file was not parsed at all and overwritten. Unfortunately all
other events would be overwritten as well. If we only want to overwrite
certain events we need to parse the file anyways.
Additionally i added a file exists check. Parsing the events from a
non-existing file does not work ;).