Removes WorldBuilder in favor of World itself
At the moment it doesn't make sense to keep WorldBuilder and World at
the same time. For the WorldBuilder::build method we would copy
everything to World and initialize. This can be done in one World.
Since we can not move the World from Python to C++ we need to create an
empty World in C++ Simulation upfront. The C++ code should be the owner
of central objects. The World is returned by reference to pybind. This
World can then be filled with geometry structures.