Refactor id tags
So far is have been modelled with 'identifiable', the implementation was
changed to allow copy of ids and to prefer composition over inheritance.

The type 'Identifiable' has been renamed to 'UniqueId' to represent that
change, further the type has become formattable and can no longer be
converted to an integer. The idea here is that you should create
interfaces that rely on a specific id type for the "thing" you want to
identify. This way you cannot mix ids of different domains.

If there are external IDs that track the progress of, for example an
agent, you need to maintain a mapping from your external id to the id
generated by jpscore.

As part of this refactoring 'special_area' lost its external_id property
since this is not part of the domain of jpscore.