Added first implementation of a neighborhood search
As optimization of the operational models the repulsion is not calculated from all other agents in the simulation, but only from the agents within a certain distance. The neighborhood search holds a grid of square cells with a fixed size, to avoid special case at the boundaries, dummy cells around the edge are included. This grid needs to be updated after each simulation step. All agents (excluding themself) in neighboring cells count as neighbors.

Note: As #72 is not merged yet the neighborhood search is based on an alias for the Level, which needs to be changed!

Co-authored-by: JetteSchumann
Signed-off-by: schroedtert <t.schroedter@fz-juelich.de>