This implements a simple Neighborhood search using a 2DGrid (#80).
The space is subdivided using a equidistant 2D Grid with an adjustable cell size. For the 2D Grid a Hash Map in two dimensions is implemented. Only the occupied Cells in the Grid are stored in the hash map. With this approach we do not need a bounding box. The Cell index in the 2D Grid can be computed by `std::floor(coordinate.x / 2)`.
The Neighborhood search uses the Grid and computes the neighborhood using the neighboring cells in the Grid. The amount of neighboring cells can be adjusted. 


Co-authored-by: JetteSchumann
Co-authored-by: schroedtert <t.schroedter@fz-juelich.de>
Co-authored-by: David Haensel <d.haensel@fz-juelich.de>